Move hex_dump out of server.c into a standalone util file
Makes it easier to port server.c wholesale - other C things still use
hex_dump
