+++ src/server.c
@@ -302,30 +302,7 @@ void server_set_debug (int n)
 	debug_client = n;
 }
 
-void hex_dump (unsigned char *p, int len) 
-{
-	int n, i, offset;
-
-	offset = 0;
-	while (len > 0) {
-	  n = len < 16 ? len : 16; 
-	  dw_printf ("  %03x: ", offset);
-	  for (i=0; i<n; i++) {
-	    dw_printf (" %02x", p[i]);
-	  }
-	  for (i=n; i<16; i++) {
-	    dw_printf ("   ");
-	  }
-	  dw_printf ("  ");
-	  for (i=0; i<n; i++) {
-	    dw_printf ("%c", isprint(p[i]) ? p[i] : '.');
-	  }
-	  dw_printf ("\n");
-	  p += 16;
-	  offset += 16;
-	  len -= 16;
-	}
-}
+void hex_dump (unsigned char *p, int len);
 
 typedef enum fromto_e { FROM_CLIENT=0, TO_CLIENT=1 } fromto_t;
